requirements.txt
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.9.0
aiohttp==0.17.2
aiohttp-jinja2==0.5.0
cached-property==1.2.0
chardet==2.3.0
fn==0.4.3
gunicorn==19.3.0
muffin==0.1.3
ujson==1.33
